feat(git-dah): add `--cooperative` option
Allow users to opt-out `--force-with-lease` and `--force-if-includes`
options from git-push command automatically invoked by git-dah.
So this is weak "Fus" shout to the full version "Fus Ro Dah".
Thinking about doing mob programming remotely, this gives extra safety
in case that multiple developers busy-pushing the same branch.

Additionally, git-dah now checks HEAD reflog when deciding to do
`pull --rebase`; which means `pull` will be skipped if the ref on the
top of remote tracking branch is included in reflog in current branch.
Merge is no longer triggered when you just amended the commit.
With `--cooperative` option, this behavior will be disabled.